Add files via upload
This includes code for module ranking, the hmm method from the paper and evaluation with MCC and AUC at the end (not sure if that code is right). The gene sets are not the gene sets used in the paper because they were not readily available I think, which is an issue, but I just found some other genes on the website linked in the comments, then used GeneCodis to get the gene sets. The merged.csv is the data obtained from Vicky's code.